gws-pipeline/pyproject.toml
[project]
name = "gws-token-activity-audit-pipeline"
version = "0.1.0"
description = "Google Workspace Token Activity Audit Pipeline"
readme = "README.md"
requires-python = ">=3.9,<3.14"
license = "MIT"
keywords = ["google workspace", "dagster", "data pipeline", "audit", "token activity", "dbt"]
authors = [{ name = "Kayvan Shah", email = "kayvan.taiyo@gmail.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
dependencies = [
    "apache-beam>=2.69.0",
    "dagster>=1.12.1",
    "google-auth>=2.43.0",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "dagster-dg-cli>=1.12.1",
    "dagster-webserver>=1.12.1",
    "pytest>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "gws_pipeline"
registry_modules = [
    "gws_pipeline.components.*",
]